# backend/internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Vibe API
  version: "1.0.0"
  description: |
    Vibe is a multi-tenant SaaS backend for padel and racket-sports court management: venue
    profiles, court and price configuration, bookings, payments, staff accounts and a public
    storefront per venue.

    Authentication is cookie based. A successful login or registration flow sets an HttpOnly
    `access_token` cookie (a 15 minute JWT) and an HttpOnly `refresh_token` cookie (a 30 day
    opaque token, scoped to `/api/v1/auth`). Every request that changes state and carries a
    session must also send the `X-CSRF-Token` header, whose value is returned in the JSON body
    of a successful login or refresh call. Public endpoints that create a session
    (register, login, refresh, logout, verify-email, resend-verification, forgot-password,
    reset-password) and the two webhook endpoints are exempt from the CSRF check because they
    either establish the session or authenticate by a provider signature instead of a cookie.

    Every JSON error response uses the envelope `{"error": <message>}`. Validation failures
    answer `422 Unprocessable Entity` with `error` as an object mapping field name to message
    or to one of the stable machine codes documented per field. Every other error status uses a
    plain string message. Paginated list endpoints share one cursor convention: an opaque
    `cursor` query parameter and a `limit` query parameter (default 50, maximum 200), with the
...

...
servers:
...
security:
  - {}
paths:
...
  /api/v1/complexes/{id}/audit-log:
    get:
      tags: [audit]
      operationId: auditListComplexLog
      summary: List a complex's own audit trail
      description: >-
        The tenant-scoped mirror of the platform audit log. The complex is taken from the
        `{id}` path segment consumed by the ownership guard, not from a query parameter, so a
        caller cannot name a tenant other than the one they own. Paginated.
      security: [{cookieAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/PathID"}
        - name: entity_type
          in: query
          schema: {type: string}
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: A page of this complex's audit log rows.
          content:
            application/json:
              schema:
                type: object
                required: [audit_logs, metadata]
                properties:
                  audit_logs:
                    type: array
                    items: {$ref: "#/components/schemas/AuditLogRow"}
                  metadata: {$ref: "#/components/schemas/Metadata"}
        "400":
          description: Invalid cursor value.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "500": {$ref: "#/components/responses/ServerError"}

  # ---------------------------------------------------------------------
  # auth
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: >-
        A 15 minute HS256 JWT set as an HttpOnly, Secure (outside development), SameSite=Lax
        cookie by login, register (in development) and refresh. Every private route resolves
        the caller from this cookie, falling back to an `Authorization: Bearer <token>` header
        when the cookie is absent.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        Required on every state-changing request that carries a session, except the routes
        listed as CSRF-exempt in the description of this document. The value is returned as
        `csrf_token` in the JSON body of a successful login or refresh call, and is validated as
        an HMAC-SHA256 of the current access token.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous page's `metadata.next_cursor`.
      schema: {type: string}
    Limit:
      name: limit
      in: query
      description: Page size. Default 50, maximum 200.
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    PathID:
      name: id
      in: path
      required: true
      description: Complex id.
      schema: {type: string, format: uuid}
...
  responses:
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Forbidden:
      description: Authenticated, but not permitted to reach this resource.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    ValidationError:
      description: One or more fields failed validation.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ValidationError"}
    RateLimited:
      description: Too many requests. See the `Retry-After` header.
      headers:
        Retry-After:
          schema: {type: integer}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    ServerError:
      description: An unexpected server error. The message is always generic, it never leaks internal detail.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human readable, safe to display.
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          description: >-
            Field name (dotted or bracketed for nested fields, e.g. `schedules[3].open_time`) to
            message. A message is either free text or one of the stable machine codes intended
            for frontend-localized copy (`slug_taken`, `deposit_percentage_over_100`,
            `deposit_exceeds_price`, `month_out_of_range`, `report_period_in_future`,
            `report_period_before_complex_existed`, `report_export_too_large`,
            `report_export_timed_out`, `price_required`).
          additionalProperties: {type: string}
    Metadata:
      type: object
      required: [has_more]
      properties:
        next_cursor: {type: string}
        has_more: {type: boolean}
        total_count: {type: integer}
...
    AuditLogRow:
      type: object
      description: Every pointer field lacks `omitempty` in Go, so an absent value serializes as JSON `null`, not as an absent key.
      required: [id, user_id, user_email, complex_id, action, entity_type, entity_id, ip_address, created_at]
      properties:
        id: {type: string, format: uuid}
        user_id: {type: [string, "null"], format: uuid}
        user_email: {type: [string, "null"], format: email}
        complex_id: {type: [string, "null"], format: uuid}
        action: {type: string}
        entity_type: {type: string}
        entity_id: {type: [string, "null"], format: uuid}
        ip_address: {type: [string, "null"]}
        created_at: {type: string, format: date-time}
...